refactored managers, remove unnecessary fields from store, and replace by loading recent details

+++ app/src/main/java/com/distributedLab/rarime/modules/rewards/RewardsClaimScreen.kt
@@ -55,7 +55,7 @@ fun RewardsClaimScreen(
 
     suspend fun submit () {
         try {
-            rewardsClaimViewModel.withdrawPoints(amountInputState.text.toDouble())
+            rewardsClaimViewModel.withdrawPoints(amountInputState.text)
 
             rewardsClaimViewModel.reloadWalletAssets()
             onBack()
